Major update: Video support, per-panel participants, key frame extraction, and UI improvements
Features:
- Video upload support with automatic key frame extraction (histogram + motion detection)
- Per-panel independent participant selection (left/right panels)
- Scanpath time-window filtering based on selected key frame interval
- Animation timestamp normalization (ms to seconds) for correct playback
- Custom data upload with column mapping and format detection
- Removed icon sidebar for cleaner UI
- Added .gitignore and requirements.txt

Files added:
- scanpath_converter.py: Scanpath parsing, normalization, key frame extraction
- upload_panel.py: Custom data upload UI with drag-and-drop
- assets/styles.css: Dashboard styling
- semantic_segmentation.py: Image segmentation utilities
- requirements.txt: Project dependencies

+++ FixaTons_methods/_compute_statistics.py
@@ -115,7 +115,7 @@ def AOI_transition_matrix(DATASET_NAME, STIMULUS_NAME, AOI, AOI_TYPE):
         scanpath = np.zeros((len(scanpath_file_lines), 5))
 
         for i in range(len(scanpath)):
-            scanpath[i, 0:4] = np.array(scanpath_file_lines[i].split()).astype(np.cfloat)
+            scanpath[i, 0:4] = np.array(scanpath_file_lines[i].split()).astype(np.complex128)
             point = Point(scanpath[i, 0], scanpath[i, 1])
             for j in range(len(polygons)):
                 if polygons[j].contains(point):
